Add examples/git_status.lua plugin
Git plugin, providing colored Git status summary information displayed in
the first part of the prompt.  Shown data include branch (colored in green if
clean, yellow if status is not clean, red if conflict is present, magenta if
clean unpublished branch, or default color if status isn't known yet) and
tracking (nothing shown if the local branch is at the same commit level as
the remote branch, otherwise the number of commits behind (↓) or ahead (↑)
of the remote are reported).
This plugin is enabled by default and can be disabled with the command
"clink set prompt.git false".