Add cdir_prompt.lua
A simple prompt filter that discards any prompt and sets it
to the errorlevel code of the last executed command (enclosed
in curly brackets), followed by the current working directory.
The the errorlevel code is colored in red if not zero.
The used color for the working directory is lightcyan.
To enable the errorlevel code (disabled by default):
    clink set prompt.errlevel true
To temporarily toggle the display of the errorlevel code, press
Ctr-Alt-E; this needs to add the following line in .inputrc:
    M-C-e: "luafunc:errlevel_change"
To disable the colored prompt: "set prompt.colored false".